+++ src/Migrations/2015_09_29_193916_create_images_table.php
@@ -10,17 +10,19 @@ class CreateImagesTable extends Migration
      */
     public function up()
     {
-        Schema::create(config('cms.db-prefix', '').'images', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('location');
-            $table->string('name')->nullable();
-            $table->string('original_name');
-            $table->string('storage_location')->default('local');
-            $table->string('alt_tag')->nullable();
-            $table->string('title_tag')->nullable();
-            $table->boolean('is_published')->default(0);
-            $table->nullableTimestamps();
-        });
+        if (in_array('images', config('cms.active-core-modules'))) {
+            Schema::create(config('cms.db-prefix', '').'images', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('location');
+                $table->string('name')->nullable();
+                $table->string('original_name');
+                $table->string('storage_location')->default('local');
+                $table->string('alt_tag')->nullable();
+                $table->string('title_tag')->nullable();
+                $table->boolean('is_published')->default(0);
+                $table->nullableTimestamps();
+            });
+        }
     }
 
     /**
@@ -28,6 +30,8 @@ class CreateImagesTable extends Migration
      */
     public function down()
     {
-        Schema::drop(config('cms.db-prefix', '').'images');
+        if (in_array('images', config('cms.active-core-modules'))) {
+            Schema::drop(config('cms.db-prefix', '').'images');
+        }
     }
 }
