+++ src/Migrations/2015_09_29_211249_create_blogs_table.php
@@ -10,17 +10,19 @@ class CreateBlogsTable extends Migration
      */
     public function up()
     {
-        Schema::create(config('cms.db-prefix', '').'blogs', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('title');
-            $table->text('entry')->nullable();
-            $table->string('tags')->nullable();
-            $table->boolean('is_published')->default(0);
-            $table->string('seo_description')->nullable();
-            $table->string('seo_keywords')->nullable();
-            $table->string('url');
-            $table->nullableTimestamps();
-        });
+        if (in_array('blog', config('cms.active-core-modules'))) {
+            Schema::create(config('cms.db-prefix', '').'blogs', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('title');
+                $table->text('entry')->nullable();
+                $table->string('tags')->nullable();
+                $table->boolean('is_published')->default(0);
+                $table->string('seo_description')->nullable();
+                $table->string('seo_keywords')->nullable();
+                $table->string('url');
+                $table->nullableTimestamps();
+            });
+        }
     }
 
     /**
@@ -28,6 +30,8 @@ class CreateBlogsTable extends Migration
      */
     public function down()
     {
-        Schema::drop(config('cms.db-prefix', '').'blogs');
+        if (in_array('blog', config('cms.active-core-modules'))) {
+            Schema::drop(config('cms.db-prefix', '').'blogs');
+        }
     }
 }
