+++ src/Migrations/2015_10_12_195952_create_faqs_table.php
@@ -10,13 +10,15 @@ class CreateFaqsTable extends Migration
      */
     public function up()
     {
-        Schema::create(config('cms.db-prefix', '').'faqs', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('question');
-            $table->text('answer');
-            $table->boolean('is_published')->default(0);
-            $table->nullableTimestamps();
-        });
+        if (in_array('faqs', config('cms.active-core-modules'))) {
+            Schema::create(config('cms.db-prefix', '').'faqs', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('question');
+                $table->text('answer');
+                $table->boolean('is_published')->default(0);
+                $table->nullableTimestamps();
+            });
+        }
     }
 
     /**
@@ -24,6 +26,8 @@ class CreateFaqsTable extends Migration
      */
     public function down()
     {
-        Schema::drop(config('cms.db-prefix', '').'faqs');
+        if (in_array('faqs', config('cms.active-core-modules'))) {
+            Schema::drop(config('cms.db-prefix', '').'faqs');
+        }
     }
 }
