+++ src/Migrations/2015_10_12_196052_create_events_table.php
@@ -10,17 +10,19 @@ class CreateEventsTable extends Migration
      */
     public function up()
     {
-        Schema::create(config('cms.db-prefix', '').'events', function (Blueprint $table) {
-            $table->increments('id');
-            $table->date('start_date');
-            $table->date('end_date');
-            $table->string('title');
-            $table->text('details');
-            $table->string('seo_description')->nullable();
-            $table->string('seo_keywords')->nullable();
-            $table->boolean('is_published')->default(0);
-            $table->nullableTimestamps();
-        });
+        if (in_array('events', config('cms.active-core-modules'))) {
+            Schema::create(config('cms.db-prefix', '').'events', function (Blueprint $table) {
+                $table->increments('id');
+                $table->date('start_date');
+                $table->date('end_date');
+                $table->string('title');
+                $table->text('details');
+                $table->string('seo_description')->nullable();
+                $table->string('seo_keywords')->nullable();
+                $table->boolean('is_published')->default(0);
+                $table->nullableTimestamps();
+            });
+        }
     }
 
     /**
@@ -28,6 +30,8 @@ class CreateEventsTable extends Migration
      */
     public function down()
     {
-        Schema::drop(config('cms.db-prefix', '').'events');
+        if (in_array('events', config('cms.active-core-modules'))) {
+            Schema::drop(config('cms.db-prefix', '').'events');
+        }
     }
 }
