+++ src/Migrations/2015_10_28_174809_create_menus_table.php
@@ -10,12 +10,14 @@ class CreateMenusTable extends Migration
      */
     public function up()
     {
-        Schema::create(config('cms.db-prefix', '').'menus', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('name');
-            $table->string('slug');
-            $table->nullableTimestamps();
-        });
+        if (in_array('menus', config('cms.active-core-modules'))) {
+            Schema::create(config('cms.db-prefix', '').'menus', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('name');
+                $table->string('slug');
+                $table->nullableTimestamps();
+            });
+        }
     }
 
     /**
@@ -23,6 +25,8 @@ class CreateMenusTable extends Migration
      */
     public function down()
     {
-        Schema::drop(config('cms.db-prefix', '').'menus');
+        if (in_array('menus', config('cms.active-core-modules'))) {
+            Schema::drop(config('cms.db-prefix', '').'menus');
+        }
     }
 }
