+++ src/Migrations/2016_02_03_175046_add_tags_to_images.php
@@ -10,9 +10,11 @@ class AddTagsToImages extends Migration
      */
     public function up()
     {
-        Schema::table(config('cms.db-prefix', '').'images', function (Blueprint $table) {
-            $table->text('tags')->nullable();
-        });
+        if (in_array('images', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'images', function (Blueprint $table) {
+                $table->text('tags')->nullable();
+            });
+        }
     }
 
     /**
@@ -20,8 +22,10 @@ class AddTagsToImages extends Migration
      */
     public function down()
     {
-        Schema::table(config('cms.db-prefix', '').'images', function ($table) {
-            $table->dropColumn('tags');
-        });
+        if (in_array('images', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'images', function ($table) {
+                $table->dropColumn('tags');
+            });
+        }
     }
 }
