+++ src/Migrations/2016_03_20_185046_add_templates_to_blogs.php
@@ -10,9 +10,11 @@ class AddTemplatesToBlogs extends Migration
      */
     public function up()
     {
-        Schema::table(config('cms.db-prefix', '').'blogs', function (Blueprint $table) {
-            $table->string('template')->default('show');
-        });
+        if (in_array('blog', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'blogs', function (Blueprint $table) {
+                $table->string('template')->default('show');
+            });
+        }
     }
 
     /**
@@ -20,8 +22,10 @@ class AddTemplatesToBlogs extends Migration
      */
     public function down()
     {
-        Schema::table(config('cms.db-prefix', '').'blogs', function ($table) {
-            $table->dropColumn('template');
-        });
+        if (in_array('blog', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'blogs', function ($table) {
+                $table->dropColumn('template');
+            });
+        }
     }
 }
