+++ src/Migrations/2016_06_01_002825_convert_to_published_at.php
@@ -10,21 +10,29 @@ class ConvertToPublishedAt extends Migration
      */
     public function up()
     {
-        Schema::table(config('cms.db-prefix', '').'pages', function (Blueprint $table) {
-            $table->dateTime('published_at')->nullable();
-        });
+        if (in_array('pages', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'pages', function (Blueprint $table) {
+                $table->dateTime('published_at')->nullable();
+            });
+        }
 
-        Schema::table(config('cms.db-prefix', '').'blogs', function (Blueprint $table) {
-            $table->dateTime('published_at')->nullable();
-        });
+        if (in_array('blog', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'blogs', function (Blueprint $table) {
+                $table->dateTime('published_at')->nullable();
+            });
+        }
 
-        Schema::table(config('cms.db-prefix', '').'events', function (Blueprint $table) {
-            $table->dateTime('published_at')->nullable();
-        });
+        if (in_array('events', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'events', function (Blueprint $table) {
+                $table->dateTime('published_at')->nullable();
+            });
+        }
 
-        Schema::table(config('cms.db-prefix', '').'faqs', function (Blueprint $table) {
-            $table->dateTime('published_at')->nullable();
-        });
+        if (in_array('faqs', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'faqs', function (Blueprint $table) {
+                $table->dateTime('published_at')->nullable();
+            });
+        }
     }
 
     /**
@@ -32,20 +40,28 @@ class ConvertToPublishedAt extends Migration
      */
     public function down()
     {
-        Schema::table(config('cms.db-prefix', '').'pages', function ($table) {
-            $table->dropColumn('published_at');
-        });
+        if (in_array('pages', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'pages', function ($table) {
+                $table->dropColumn('published_at');
+            });
+        }
 
-        Schema::table(config('cms.db-prefix', '').'blogs', function ($table) {
-            $table->dropColumn('published_at');
-        });
+        if (in_array('blog', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'blogs', function ($table) {
+                $table->dropColumn('published_at');
+            });
+        }
 
-        Schema::table(config('cms.db-prefix', '').'events', function ($table) {
-            $table->dropColumn('published_at');
-        });
+        if (in_array('events', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'events', function ($table) {
+                $table->dropColumn('published_at');
+            });
+        }
 
-        Schema::table(config('cms.db-prefix', '').'faqs', function ($table) {
-            $table->dropColumn('published_at');
-        });
+        if (in_array('faqs', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'faqs', function ($table) {
+                $table->dropColumn('published_at');
+            });
+        }
     }
 }
