+++ src/Migrations/2017_06_10_234231_add_order_to_menus.php
@@ -11,9 +11,11 @@ class AddOrderToMenus extends Migration
      */
     public function up()
     {
-        Schema::table(config('cms.db-prefix', '').'menus', function (Blueprint $table) {
-            $table->text('order')->nullable();
-        });
+        if (in_array('menus', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'menus', function (Blueprint $table) {
+                $table->text('order')->nullable();
+            });
+        }
     }
 
     /**
@@ -21,8 +23,10 @@ class AddOrderToMenus extends Migration
      */
     public function down()
     {
-        Schema::table(config('cms.db-prefix', '').'menus', function ($table) {
-            $table->dropColumn('order');
-        });
+        if (in_array('menus', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'menus', function ($table) {
+                $table->dropColumn('order');
+            });
+        }
     }
 }
