+++ src/Migrations/2017_12_14_045216_add_hero_images.php
@@ -13,13 +13,17 @@ class AddHeroImages extends Migration
      */
     public function up()
     {
-        Schema::table(config('cms.db-prefix', '').'blogs', function (Blueprint $table) {
-            $table->string('hero_image')->nullable();
-        });
+        if (in_array('blog', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'blogs', function (Blueprint $table) {
+                $table->string('hero_image')->nullable();
+            });
+        }
 
-        Schema::table(config('cms.db-prefix', '').'pages', function (Blueprint $table) {
-            $table->string('hero_image')->nullable();
-        });
+        if (in_array('pages', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'pages', function (Blueprint $table) {
+                $table->string('hero_image')->nullable();
+            });
+        }
     }
 
     /**
@@ -29,12 +33,16 @@ class AddHeroImages extends Migration
      */
     public function down()
     {
-        Schema::table(config('cms.db-prefix', '').'blogs', function ($table) {
-            $table->dropColumn('hero_image');
-        });
+        if (in_array('blog', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'blogs', function ($table) {
+                $table->dropColumn('hero_image');
+            });
+        }
 
-        Schema::table(config('cms.db-prefix', '').'pages', function ($table) {
-            $table->dropColumn('hero_image');
-        });
+        if (in_array('pages', config('cms.active-core-modules'))) {
+            Schema::table(config('cms.db-prefix', '').'pages', function ($table) {
+                $table->dropColumn('hero_image');
+            });
+        }
     }
 }
